Start work on component paradigms.
- Added ArraySlice class to provide mutable views into unowned memory
- Added ComponentParadigms, which represent a series of
components that an entity could be defined by. These
paradigms are then stored in possibly many memory blocks
where the contents of the blocks are SoA's.
- Fixed an issue where the Component key set was not getting
calculated properly.
- `ComponentId` is now called `ComponentKey`
- `ComponentKey` overall size has been changed to 4 bytes
instead of 8
- `ComponentKey` is now composed of 3 fields.
`m_FullComponentKey`, `m_Key` which represents the actual
component key mask, and `m_Set` which represents which
component set the component key is a part of.